requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
python-telegram-bot==20.4
tqdm==4.65.0
cloudscraper==1.2.71
